verify phone number on feed Profile Completion

+++ app/src/main/java/com/collegedekho/app/resource/ScrollingFABBehavior.java
@@ -31,7 +31,7 @@ public class ScrollingFABBehavior extends CoordinatorLayout.Behavior<FloatingAct
                 CoordinatorLayout.LayoutParams lp = (CoordinatorLayout.LayoutParams) fab.getLayoutParams();
                // int fabBottomMargin = lp.bottomMargin;
                 int distanceToScroll = fab.getHeight() + mFabMenuMargin;
-                float ratio = (float)dependency.getY()/(float)toolbarHeight;
+                float ratio = dependency.getY()/(float)toolbarHeight;
                 fab.setTranslationY(-distanceToScroll * ratio);
         }
         return true;
